requirements.txt
pandas
sqlalchemy
sqlalchemy_utils
requests
dbt-core
dbt-postgres
psycopg2-binary
streamlit
folium
streamlit_folium